Refactored getSelectedFacilityFunctions into separate service class.
